% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbreviate_states.R
\name{abbreviate_states}
\alias{abbreviate_states}
\title{Abbreviate states}
\usage{
abbreviate_states(state_names)
}
\arguments{
\item{state_names}{\code{character} of state names in full or abbreviated form}
}
\value{
\code{character} of state names in abbreviated forms
}
\description{
Abbreviate states
}
\examples{
abbreviate_states(
  c(
    "Victoria",
    "Tasmania",
    "VIC"
  )
)

}
