% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviour.change.analysis-package.R
\docType{package}
\name{behaviour.change.analysis-package}
\alias{behaviour.change.analysis}
\alias{behaviour.change.analysis-package}
\title{behaviour.change.analysis: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
