% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_survey_gam.R
\name{fit_survey_gam}
\alias{fit_survey_gam}
\title{Fit survey gam}
\usage{
fit_survey_gam(fit_dat, pred_dat)
}
\arguments{
\item{fit_dat}{\code{data.frame}}

\item{pred_dat}{\code{data.frame}}
}
\description{
Fit survey gam
}
\details{
Fits and predicts a gam over survey data
}
