% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_smooth.R
\name{gaussian_smooth}
\alias{gaussian_smooth}
\title{Gaussian smooth for data}
\usage{
gaussian_smooth(values, sd = 1, ...)
}
\arguments{
\item{values}{\code{numeric}}

\item{sd}{\code{numeric}}

\item{...}{}
}
\description{
Gaussian smooth for data
}
