% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday_dates.R
\name{holiday_dates}
\alias{holiday_dates}
\title{Australian holiday dates}
\usage{
holiday_dates(holiday_file = "data/holidays/public_holidays.csv")
}
\arguments{
\item{holiday_file}{\code{character}}
}
\description{
Dates of Australian public holidays in 2020—2023 by state, from:
\url{https://data.gov.au/dataset/ds-dga-b1bc6077-dadd-4f61-9f8c-002ab2cdff10/details?q=}
}
\examples{

holiday_dates(holiday_file = tempfile())

}
