% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interventions.R
\name{interventions}
\alias{interventions}
\title{Interventions}
\usage{
interventions(
  which = c("all", "national", "act", "nsw", "nt", "qld", "sa", "tas", "vic", "wa"),
  end_dates = FALSE,
  exclude_after = NA
)
}
\arguments{
\item{which}{\code{character}}

\item{exclude_after}{}
}
\value{
\code{tbl_df}
}
\description{
Start and (optionally) end dates for governmental non-pharmaceutical intervention actions in Australia nationally and at state level from 2020—2021.
}
\examples{
interventions()
}
