% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_ggplot.R
\name{save_ggplot}
\alias{save_ggplot}
\title{Save ggplot}
\usage{
save_ggplot(
  filename,
  dir = "outputs",
  subdir = "figures",
  multi = TRUE,
  width = 11.69/2,
  height = 8.27/3,
  scale = 1,
  dpi = 150
)
}
\arguments{
\item{dpi}{}
}
\description{
Save ggplot
}
\details{
Saves a ggplot into a specified dir with a specified aspect ratio
}
