% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_holiday_dates.R
\name{school_holiday_dates}
\alias{school_holiday_dates}
\title{Australian School Holiday Dates 2020—2022}
\usage{
school_holiday_dates()
}
\value{
\code{tbl_df}
}
\description{
Australian School Holiday Dates 2020—2022
}
\details{
dates of school holidays by state, from:
2020: \url{https://info.australia.gov.au/about-australia/special-dates-and-events/school-term-dates}
2021 onwards:
\url{https://www.education.act.gov.au/__data/assets/pdf_file/0007/801583/Term-Date-2019-2025.pdf}
\url{https://education.nsw.gov.au/public-schools/going-to-a-public-school/calendars}
\url{https://nt.gov.au/learning/primary-and-secondary-students/school-term-dates-in-nt}
\url{https://education.qld.gov.au/about-us/calendar/term-dates}
\url{https://www.education.sa.gov.au/teaching/south-australian-state-schools-term-dates}
\url{https://www.education.tas.gov.au/about-us/term-dates/term-dates-2021/}
<https://www.education.wa.edu.au/future-term-dates/#' >
\url{https://www.education.vic.gov.au/about/department/Pages/datesterm.aspx}
}
\examples{
school_holiday_dates()
}
