% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unabbreviate_states.R
\name{unabbreviate_states}
\alias{unabbreviate_states}
\title{Unabbreviate states}
\usage{
unabbreviate_states(state_names)
}
\arguments{
\item{state_names}{\code{character} of state names in full or abbreviated form}
}
\value{
\code{character} of state names in full
}
\description{
Unabbreviate states
}
\examples{
unabbreviate_states(
  c(
    "Victoria",
    "Tasmania",
    "VIC"
  )
)

}
