% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_ticks_and_labels.R
\name{split_ticks_and_labels}
\alias{split_ticks_and_labels}
\title{Split ticks and labels}
\usage{
split_ticks_and_labels(
  dat = NULL,
  tick_freq = "1 month",
  label_freq = "2 months",
  label_format = "\%b \%y",
  label_shift = TRUE,
  start_date,
  end_date
)
}
\arguments{
\item{label_shift}{\code{logical} sometimes necessary or desirable to switch which
ticks are labelled (e.g. to align with start of a year etc.)}

\item{end_date}{}
}
\value{
\code{list} of tick dates, labels, and tick colours
}
\description{
Independently split frequency of plotting labels and ticks
}
\examples{
split_ticks_and_labels(
tick_freq = "3 month",
label_freq = "6 months",
label_format = "\%b \%y"
start_date = as.Date("2020-01-01"),
end_date = as.Date("2022-12-31"),
)

}
